Fized previouse issues with settings
This version gives up to 100ms faster starting times.

+++ src/settings.py
@@ -35,6 +35,17 @@ if open(currentProfile, "r").readline().replace("\n", "") == "":
         f.write("profiles/settings.json")
     print("Profile variable was empty, set it to settings.json")
 
+profile = open(currentProfile, "r").readline().replace("\n", "")
+if not os.path.exists(profile) or open(profile, "r").readline() == "":
+    with open(profile, "w") as f:
+        f.truncate(0)
+        f.write("{}")
+        f.close()
+        print("Settings file didn't exist, created a new one")
+
+settings = json.load(open(profile, "r"))
+settings_file = open(profile, "w")
+
 # Check for settings file in root folder
 SETPATH = str(os.path.abspath(os.path.join(PATH, os.pardir)))
 # Check that settings.json exists
@@ -53,12 +64,10 @@ def EnsureFile(file:str):
     Args:
         file (str): Filename.
     """
-    try:
-        with open(file, "r") as f:
-            pass
-    except:
-        with open(file, "w") as f:
-            f.write("{}")
+    if os.path.exists(file):
+        return
+    with open(file, "w") as f:
+        f.truncate(0)
 
 def ChangeProfile():
     """Will change the currently selected profile and reload the app.
@@ -79,10 +88,7 @@ def CreateProfile():
     """
     from tkinter import filedialog
     newFile = filedialog.asksaveasfile(initialdir=PATH+"\\profiles", initialfile="newProfile.json", title="Create a new profile", filetypes=(("JSON files", "*.json"), ("All files", "*.*")))
-    try:       
-        # Copy the current profile
-        profile = open(currentProfile, "r").readline().replace("\n", "")
-            
+    try: 
         with open(profile, "r") as f:
             data = json.load(f)
         
@@ -92,10 +98,7 @@ def CreateProfile():
         filePath = newFile.name
         
         newFile.close()
-            
-        # Change the current profile
-        # ChangeProfile(filePath)
-    
+
     except Exception as ex:
         print(ex.args)
         print("Failed to create profile")
@@ -112,15 +115,8 @@ def UpdateSettings(category:str, name:str, data:any):
     """
     global currentFrameSettings, frameCounter
     try:
-        profile = open(currentProfile, "r").readline().replace("\n", "")
-        EnsureFile(profile)
-        with open(profile, "r") as f:
-            settings = json.load(f)
-
         settings[category][name] = data
-        with open(profile, "w") as f:
-            f.truncate(0)
-            json.dump(settings, f, indent=6)
+        json.dump(settings, settings_file, indent=6)
             
         currentFrameSettings = settings
         frameCounter = FRAMECOUNTER
@@ -142,22 +138,12 @@ def GetSettings(category:str, name:str, value:any=None):
     Returns:
         _type_: The data from the json file. (or the default value)
     """
-    global frameCounter, currentFrameSettings
-    try:
-        if frameCounter != FRAMECOUNTER:
-            profile = open(currentProfile, "r").readline().replace("\n", "")
-            EnsureFile(profile)
-            with open(profile, "r") as f:
-                settings = json.load(f)
-                currentFrameSettings = settings
-            frameCounter = FRAMECOUNTER
-        else:
-            settings = currentFrameSettings
         
+    try:
         if settings[category][name] == None:
             return value    
-        
-        return settings[category][name]
+        else:
+            return settings[category][name]
     except Exception as ex:
         if value != None:
             CreateSettings(category, name, value)
@@ -177,11 +163,6 @@ def CreateSettings(category:str, name:str, data:any):
     """
     global currentFrameSettings, frameCounter
     try:
-        profile = open(currentProfile, "r").readline().replace("\n", "")
-        EnsureFile(profile)
-        with open(profile, "r") as f:
-            settings = json.load(f)
-
         # If the setting doesn't exist then create it 
         if not category in settings:
             settings[category] = {}
@@ -211,15 +192,11 @@ def AddToList(category:str, name:str, data:any, exclusive:bool=False):
     """
     global currentFrameSettings, frameCounter
     try:
-        profile = open(currentProfile, "r").readline().replace("\n", "")
-        EnsureFile(profile)
-        with open(profile, "r") as f:
-            settings = json.load(f)
-
         # If the setting doesn't exist then create it 
         if not category in settings:
             settings[category] = {}
             settings[category][name] = []
+
             # Check if the data is a list
             if isinstance(data, list):
                 for item in data:
@@ -273,11 +250,6 @@ def RemoveFromList(category:str, name:str, data:any):
     """
     global currentFrameSettings, frameCounter
     try:
-        profile = open(currentProfile, "r").readline().replace("\n", "")
-        EnsureFile(profile)
-        with open(profile, "r") as f:
-            settings = json.load(f)
-
         # If the setting doesn't exist then don't do anything 
         if not category in settings:
             return
